Modlist revision and more code cleanup and file restructuring.
