+++ all-the-overhaul-modpack/data.lua
@@ -13,12 +13,11 @@ end
 Test(true)
 
 --call all needed scripts
-require("prototypes/item/add")
-require("prototypes/recipe/add")
-require("prototypes/entity/imersium-wall")
+require("prototypes/imersium-wall")
 require("prototypes/common")
 require("prototypes/pipes")
 require("prototypes/bio-industries")
+require("prototypes/technology")
 if mods["RampantArsenal"] then
     require("prototypes/rampant-arsenal")
 end
