--- all-the-overhaul-modpack/prototypes/entity/fu_magnet_entity.lua
@@ -1,1 +1,0 @@
-data.raw["assembling-machine"]["fu_magnet_entity"].ingredient_count = 3
