--- all-the-overhaul-modpack/prototypes/item/add.lua
@@ -1,14 +1,0 @@
---Add new items
-data:extend({
-  {
-    type = "item",
-    name = "imersium-wall",
-    icon = "__all-the-overhaul-modpack__/graphics/icons/imersium-wall.png",
-    icon_size = 64,
-    icon_mipmaps = 1,
-    order = "a[stone-wall]a[steel-wall]-c[imersium-wall]",
-    place_result = "imersium-wall",
-    stack_size = 50,
-    subgroup = "defensive-structure",
-  }
-})
